renamed pid_sim, fix issue with config types, upd issue with not connected state transitions
